+++ src/element/basic_shapes.rs
@@ -158,7 +158,8 @@ impl<Coord, DB: DrawingBackend> Drawable<DB> for Rectangle<Coord> {
         backend: &mut DB,
     ) -> Result<(), DrawingErrorKind<DB::ErrorType>> {
         match (points.next(), points.next()) {
-            (Some(mut a), Some(mut b)) => {
+            (Some(a), Some(b)) => {
+                let (mut a, mut b) = ((a.0.min(b.0), a.1.min(b.1)), (a.0.max(b.0), a.1.max(b.1)));
                 a.1 += self.margin.0 as i32;
                 b.1 -= self.margin.1 as i32;
                 a.0 += self.margin.2 as i32;